Sprint 2 - Migrating test classes for login and validation to individual files.
