+++ tests/test_sprint1.py
@@ -103,226 +103,6 @@ class TestApplicationMethods(TestCase):
         self.assertTrue(server_response['status'] == "success")
 
 
- def get_dict_from_response(response):
-        return response.json()
-
-
-class TestLogin(TestCase):
-    def setUp(self):
-        self.client = Client()
-        self.user =   Userinfo.objects.create(username = "nikhil_test", password = "Welcome@2018", emailID = "nikhil_test@gmail.com")
-        self.valid_token_user = jwt.decode({'identifier': self.user.username, 'password': self.user.password})
-        self.valid_token_email = jwt.decode({'identifier': self.user.username, 'password': self.user.password})
-
-    def test_login_valid_username(self):
-        username = self.user.username
-        password = self.user.password
-        response = self.client.post('api/v1/validuser/', data={'identifier': username, 'password': password})
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('token' in pay_load.keys())
-        auth_token = pay_load['token']
-
-        #Validate the authorization token signature.
-        pay_load = jwt.decode(auth_token, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-        self.assertTrue(pay_load["identifier"] == username)
-        self.assertTrue(pay_load["password"] == password)
-
-    def test_login_valid_email(self):
-        emaiId = self.user.emailID
-        password = self.user.password
-        response = self.client.post('api/v1/validuser/', data={'identifier': emaiId, 'password': password})
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('token' in pay_load.keys())
-        auth_token = pay_load['token']
-
-        #Validate the authorization token signature.
-        pay_load = jwt.decode(auth_token, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-        self.assertTrue(pay_load["identifier"] == emaiId)
-        self.assertTrue(pay_load["password"] == password)
-
-        #validate status code
-        self.assertTrue(response.status_code == 200)
-
-    def test_login_invalid_username(self):
-        username = "nikhil112233"
-        password = self.user.password
-        response = self.client.post('api/v1/validuser/', data={'identifier': username, 'password': password})
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('errors' in pay_load.keys())
-        errors_dict = pay_load['token']
-        self.assertTrue(type(errors_dict) == dict)
-        self.assertTrue('form' in errors_dict.keys())
-        error_message = errors_dict['form']
-        self.assertTrue(error_message == 'Invalid Credentials')
-
-        # validate status code
-        self.assertTrue(response.status_code == 401)
-
-        # validate status code
-        self.assertTrue(response.status_code == 401)
-
-    def test_login_invalid_email(self):
-        emaiId = "nikhil112233@gmail.com"
-        password = self.user.password
-        response = self.client.post('api/v1/validuser/', data={'identifier': emaiId, 'password': password})
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('errors' in pay_load.keys())
-        errors_dict = pay_load['token']
-        self.assertTrue(type(errors_dict) == dict)
-        self.assertTrue('form' in errors_dict.keys())
-        error_message = errors_dict['form']
-        self.assertTrue(error_message == 'Invalid Credentials')
-
-        # validate status code
-        self.assertTrue(response.status_code == 401)
-
-
-    def test_login_invalid_password(self):
-        username = self.user.username
-        password = 'nikhilaabbcc'
-        response = self.client.post('api/v1/validuser/', data={'identifier': username, 'password': password})
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('errors' in pay_load.keys())
-        errors_dict = pay_load['token']
-        self.assertTrue(type(errors_dict) == dict)
-        self.assertTrue('form' in errors_dict.keys())
-        error_message = errors_dict['form']
-        self.assertTrue(error_message == 'Invalid Credentials')
-
-        # validate status code
-        self.assertTrue(response.status_code == 401)
-
-class TestIsRequestValid(TestCase):
-    def setUp(self):
-        self.client = Client()
-        self.user = Userinfo.objects.create(username="nikhil_test", password="Welcome@2018",
-                                            emailID="nikhil_test@gmail.com")
-        self.valid_token = jwt.encode({'username': self.user.username}, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-
-    #Missing Authentication token
-    def test_missing_auth_request(self):
-        response = self.client.post('http://localhost:8000/api/v1/validuser/')
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('isValid' in pay_load.keys())
-        flag = pay_load['isValid']
-        self.assertTrue(flag == False)
-
-    #has a Invalid token
-    def test_invalid_auth_request(self):
-        self.client.defaults['HTTP_AUTHORIZATION'] = "Nikhil.Invalid.User.TryingtoHack"
-        response = self.client.post('http://localhost:8000/api/v1/validuser/')
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('isValid' in pay_load.keys())
-        flag = pay_load['isValid']
-        self.assertTrue(flag == False)
-        self.client.defaults = {}
-
-
-    #has a Valid token
-    def test_valid_auth_request(self):
-        self.client.defaults['HTTP_AUTHORIZATION'] = self.valid_token
-        response = self.client.post('http://localhost:8000/api/v1/validuser/')
-        pay_load = get_dict_from_response(response)
-        self.assertTrue('isValid' in pay_load.keys())
-        flag = pay_load['isValid']
-        self.assertTrue(flag == True)
-        self.client.defaults = {}
-
-
-class TestValidate(TestCase):
-
-    def setUp(self):
-        self.client = Client()
-        self.user = Userinfo.objects.create(username="nikhil_test", password="Welcome@2018",
-                                            emailID="nikhil_test@gmail.com")
-        self.valid_token = jwt.encode({'username': self.user.username}, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-
-
-    def test_is_user_valid_with_missing_token(self):
-        flag, message = Validate.is_user_valid()
-        self.assertTrue(flag == False)
-        self.assertTrue(message == 'Missing Token')
-
-    def test_is_user_valid_with_empty_token(self):
-        flag, message = Validate.is_user_valid('')
-        self.assertTrue(flag == False)
-        self.assertTrue(message == 'Missing Token')
-
-    def test_is_user_valid_with_invalid_token(self):
-        token = "Nikhil.Invalid.User.TryingtoHack"
-        flag, message = Validate.is_user_valid(token)
-        self.assertTrue(flag == False)
-        self.assertTrue(message == 'Invalid User')
-
-    def test_is_user_valid_with_valid_token(self):
-        token = self.valid_token
-        flag, message = Validate.is_user_valid(token)
-        self.assertTrue(flag == True)
-        self.assertTrue(message == 'Valid User')
-
-
-
-class TestAuthenticate(TestCase):
-    def setUp(self):
-        self.user = Userinfo.objects.create(username="nikhil_test", password="Welcome@2018",
-                                            emailID="nikhil_test@gmail.com")
-
-    def test_authenticate_user_with_valid_username(self):
-        flag, message, uname = Authenticate.authenticate_user(self.user.username, self.user.password)
-        self.assertTrue(flag == True)
-        self.assertTrue(message = "User Authenticated")
-        self.assertTrue(uname == self.user.username)
-
-    def test_authenticate_user_with_valid_email(self):
-        flag, message, uname = Authenticate.authenticate_user(self.user.emailID, self.user.password)
-        self.assertTrue(flag == True)
-        self.assertTrue(message = "User Authenticated")
-        self.assertTrue(uname == self.user.username)
-
-    def test_authenticate_user_with_invalid_username(self):
-        flag, message, uname = Authenticate.authenticate_user("Nikhil567", self.user.password)
-        self.assertTrue(flag == False)
-        self.assertTrue(message = "Invalid Credentials")
-        self.assertTrue(uname == '')
-
-    def test_authenticate_user_with_invalid_email(self):
-        flag, message, uname= Authenticate.authenticate_user("Nikhil910@gmail.com", self.user.password)
-        self.assertTrue(flag == False)
-        self.assertTrue(message = "Invalid Credentials")
-        self.assertTrue(uname == '')
-
-    def test_authenticate_user_with_invalid_password(self):
-        flag, message, flag,uname = Authenticate.authenticate_user(self.user.username, "Nikhilabc")
-        self.assertTrue(flag == False)
-        self.assertTrue(message = "Invalid Credentials")
-        self.assertTrue(uname == '')
-
-        flag, message, uname = Authenticate.authenticate_user(self.user.emailID, "Nikhilabc")
-        self.assertTrue(flag == False)
-        self.assertTrue(message="Invalid Credentials")
-        self.assertTrue(uname == '')
-
-    def test_generate_token_valid_payload(self):
-        pay_load = {"username": self.user.username}
-        token = Authenticate.generate_token(pay_load)
-        self.assertTrue(token is not None and token is not '')
-        pay_load_returned = jwt.decode(token, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-
-        self.assertTrue(pay_load_returned['username'] == pay_load['username'])
-
-    def test_generate_token_id_valid_username(self):
-        token = Authenticate.generate_token_id(self.user.username)
-        pay_load_returned = jwt.decode(token, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-
-        self.assertTrue(pay_load_returned['username'] == self.user.username)
-
-    def test_generate_token_id_valid_email(self):
-        token = Authenticate.generate_token_id(self.user.emailID)
-        pay_load_returned = jwt.decode(token, 'ThisU$erI$LoggedInBetoInfo', algorithm="HS256")
-
-        self.assertTrue(pay_load_returned['username'] == self.user.username)
-
-
