+++ tests/utils.py
@@ -1,0 +1,2 @@
+def get_dict_from_response(response):
+    return response.json()
